+++ CHANGES.md
@@ -1,3 +1,7 @@
+## Version 8.6.0
+
+Unreleased
+
 ## Version 8.5.1
 
 Unreleased